Recommend pipewire-module-xrdp
Allow xrdp users to have audio working out-of-the-box if ever pipewire
is installed. It works via a XDG autostart script that loads the xrdp
pipewire modules when the session is started.

If ever pipewire is not running, the XDG autostart script does nothing,
thus it's safe to have pipewire-module-xrdp installed on a system that
uses pulseaudio, or that doesn't have any audio stack at all.

pipewire-module-xrdp is developed and maintained by the same upstream as
xrdp, at: https://github.com/neutrinolabs/pipewire-module-xrdp.

The Debian package has just been released:
https://tracker.debian.org/pkg/pipewire-module-xrdp